agent-scaffolding/skills/ml-engineer/assets/requirements.txt
# =============================================================================
# Mage + H2O Python Dependencies
# =============================================================================
# This requirements.txt should be mounted or baked into the Mage container
# to enable communication with the H2O cluster.
# =============================================================================

# H2O Python Client - REQUIRED for h2o.init() from Mage blocks
h2o>=3.46.0

# Data processing
pandas>=2.0.0
polars>=0.20.0
numpy>=1.24.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# API/HTTP clients
requests>=2.31.0
httpx>=0.25.0

# Object storage
boto3>=1.34.0  # AWS S3
google-cloud-storage>=2.14.0  # GCS

# Observability
prometheus-client>=0.19.0

# Alerting
slack-sdk>=3.26.0
